% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.FATES.R
\name{met2model.FATES}
\alias{met2model.FATES}
\title{met2model for FATES}
\usage{
met2model.FATES(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  lst = 0,
  lat,
  lon,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded}

\item{end_date}{the end date of the data to be downloaded}

\item{lst}{timezone offset to GMT in hours}

\item{lat, lon}{latitude and longitude of site in decimal degrees}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose for(year in start_year:end_year)}

\item{...}{additional arguments, currently ignored}
}
\description{
met2model wrapper for FATES
}
