% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.FATES.R
\name{model2netcdf.FATES}
\alias{model2netcdf.FATES}
\title{Code to convert FATES netcdf output into into CF standard}
\usage{
model2netcdf.FATES(outdir)
}
\arguments{
\item{outdir}{Location of FATES model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\author{
Michael Dietze
}

