% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.FATES.R
\name{model2netcdf.FATES}
\alias{model2netcdf.FATES}
\title{Code to convert FATES netcdf output into into CF standard}
\usage{
model2netcdf.FATES(outdir)
}
\arguments{
\item{outdir}{Location of FATES model output}
}
\description{
Code to convert FATES netcdf output into into CF standard
}
\examples{
 
\dontrun{
example.output <- system.file("case.clm2.h0.2004-01-01-00000.nc",package="PEcAn.FATES")
model2netcdf.FATES(outdir="~/")
}

}
\author{
Michael Dietze, Shawn Serbin
}
